Reworked view structure
