+++ mage_game/model/character.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
-from .magic import Element, Projectile, SpellAttribute
+from .magic import *
 
 
 @dataclass
