+++ mage_game/view/__init__.py
@@ -1,0 +1,2 @@
+from .graphics import *
+from .sprites import *
