+++ test/test_projectile.py
@@ -1,7 +1,7 @@
-from mage_game.model import Element, Projectile, SpellAttribute
+from mage_game.model import Element, ProjectileSprite, SpellAttribute
 
 
 def test_projectile_default():
-    default = Projectile()
+    default = ProjectileSprite()
     assert default.element() == Element.NONE
     assert default.get_attribute(SpellAttribute.DAMAGE) == 1
